requirements.txt
   streamlit
   openai
   pandas
   python-dotenv